Fixed Issue with Reusable Views
Item views that were added after the current visible amount of item views were not being set properly. If a view isn't null, that means that it contains a layout and can be reused. This indicates that we can find the views id's within the layout in order to set them to the updated task information. The previous code only referenced these id's when the view was null. We also need to reference these same id's when the views are not null, indicating that the layout has been created previously and needs to be reused.